+++ composer.json
@@ -6,7 +6,7 @@
   "license": "GPL-3.0-only",
   "require": {
     "oxid-esales/oxideshop-ce": "^v6.3",
-    "php": ">=7.1"
+    "php": ">=7.3"
   },
   "require-dev": {
   },
