--- docs/uninstall.sql
@@ -1,5 +1,0 @@
---
--- Remove table `oxpspasswordpolicy_attempt`
---
-
-DROP TABLE `oxpspasswordpolicy_attempt`
