--- src/Controller/ControllerWithPasswordPolicy.php
@@ -1,107 +1,0 @@
-<?php
-
-/**
- * This file is part of OXID Professional Services Password Policy module.
- *
- * OXID Professional Services Password Policy module is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * OXID Professional Services Password Policy module is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with OXID Professional Services Password Policy module.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @author        OXID Professional services
- * @link          https://www.oxid-esales.com
- * @copyright (C) OXID eSales AG 2003-2021
- */
-
-namespace OxidProfessionalServices\PasswordPolicy\Controller;
-
-use OxidProfessionalServices\PasswordPolicy\Core\PasswordPolicyModule;
-
-/**
- * Class oxpsPasswordPolicyUser
- */
-trait ControllerWithPasswordPolicy
-{
-
-    /**
-     * @var object $passwordPolicy Password policy module instance.
-     */
-    protected $passwordPolicy;
-
-    /**
-     * Overridden init method, that creates password policy module object.
-     */
-    public function init()
-    {
-
-        // Parent call
-        $this->oxpsPasswordPolicyInitParent();
-
-        $this->setPasswordPolicy();
-    }
-
-    /**
-     * Parent `init` call. Method required for mocking.
-     *
-     * @return mixed
-     */
-    protected function oxpsPasswordPolicyInitParent()
-    {
-        // @codeCoverageIgnoreStart
-        return parent::init();
-        // @codeCoverageIgnoreEnd
-    }
-
-    /**
-     * Set Password Policy instance
-     *
-     * @param mixed $mPasswordPolicy
-     */
-    public function setPasswordPolicy($oPasswordPolicy = null)
-    {
-        $this->passwordPolicy = is_object($oPasswordPolicy) ? $oPasswordPolicy : oxNew(PasswordPolicyModule::class);
-    }
-
-    /**
-     * Overridden registration render method to add password policy parameters.
-     *
-     * @return string
-     */
-    public function render()
-    {
-
-        // Assign current settings values
-        $this->_aViewData = array_merge($this->_aViewData, $this->getPasswordPolicy()->getModuleSettings());
-
-        // Parent call
-        return $this->oxpsPasswordPolicyRenderParent();
-    }
-
-    /**
-     * @return object Password policy module instance.
-     */
-    public function getPasswordPolicy()
-    {
-        return $this->passwordPolicy;
-    }
-
-    /**
-     * Parent `render` call. Method required for mocking.
-     *
-     * @return mixed
-     */
-    protected function oxpsPasswordPolicyRenderParent()
-    {
-        // @codeCoverageIgnoreStart
-        return parent::render();
-        // @codeCoverageIgnoreEnd
-    }
-}
