--- tests/unit/Controller/RegisterControllerTest.php
@@ -1,91 +1,0 @@
-<?php
-
-/**
- * This file is part of OXID Professional Services Password Policy module.
- *
- * OXID Professional Services Password Policy module is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * OXID Professional Services Password Policy module is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with OXID Professional Services Password Policy module.  If not, see <http://www.gnu.org/licenses/>.
- *
- * @author        OXID Professional services
- * @link          http://www.oxid-esales.com
- * @copyright (C) OXID eSales AG 2003-2019
- */
-
-/**
- * Tests for "OxpsPasswordPolicyRegister" class
- */
-class Unit_Module_Controllers_OxpsPasswordPolicyRegisterTest extends OxidTestCase
-{
-
-    /**
-     * @var OxpsPasswordPolicyRegister
-     */
-    protected $SUT;
-
-
-    /**
-     * Set initial objects state.
-     *
-     * @return null|void
-     */
-    public function setUp()
-    {
-        parent::setUp();
-
-        $this->SUT = $this->getMock('OxpsPasswordPolicyRegister', array(
-            '_oxpsPasswordPolicyRegister_init_parent',
-            '_oxpsPasswordPolicyRegister_render_parent',
-        ));
-    }
-
-
-    /**
-     * `getPasswordPolicy` returns null if init have not run (nothing was set)
-     */
-    public function testGetPasswordPolicy_initNotRan_returnNull()
-    {
-        $this->assertNull($this->SUT->getPasswordPolicy());
-    }
-
-    /**
-     * `getPasswordPolicy` returns `PasswordPolicyModule` object after init.
-     */
-    public function testGetPasswordPolicy_initRan_returnPasswordPolicyModuleInstance()
-    {
-        $this->SUT->init();
-
-        $this->assertType('OxpsPasswordPolicyModule', $this->SUT->getPasswordPolicy());
-    }
-
-
-    /**
-     * `render` should add Password Policy settings to ViewData array and call parent
-     */
-    public function testRender_addPasswordPolicySettingsCallParent()
-    {
-        $this->SUT->expects($this->once())->method('_oxpsPasswordPolicyRegister_render_parent');
-
-        $aConfigKeys = array(
-            'iMaxAttemptsAllowed', 'iTrackingPeriod', 'blAllowUnblock',
-            'iMinPasswordLength', 'iGoodPasswordLength', 'iMaxPasswordLength', 'aPasswordRequirements'
-        );
-
-        $this->SUT->init();
-        $this->SUT->render();
-        $aViewData = $this->SUT->getViewData();
-
-        foreach ($aConfigKeys as $sKey) {
-            $this->assertArrayHasKey($sKey, $aViewData);
-        }
-    }
-}
