+++ views/admin/de/passwordpolicy_lang.php
@@ -21,13 +21,6 @@ $aLang = array(
     'charset' => 'UTF-8',
     'oxpspasswordpolicy' => 'Password Policy',
     'OXPS_PASSWORDPOLICY_ADMIN_TITLE' => 'Password Policy Einstellungen',
-    'OXPS_PASSWORDPOLICY_ADMIN_ACCOUNTBLOCK' => 'Automatisches Blockieren des Accounts',
-    'OXPS_PASSWORDPOLICY_ADMIN_MAXATTEMPTS' => 'Maximale Anmeldeversuche',
-    'OXPS_PASSWORDPOLICY_ADMIN_MAXATTEMPTS_HINT' => 'Max. Anzahl der Anmeldeversuche, bevor ein Benutzer blockiert wird.',
-    'OXPS_PASSWORDPOLICY_ADMIN_TRACKINGPERIOD' => 'Überwachungszeitraum',
-    'OXPS_PASSWORDPOLICY_ADMIN_TRACKINGPERIOD_HINT' => 'Zeitraum in Minuten, in dem nicht erfolgreiche Anmeldeversuche überwacht werden.',
-    'OXPS_PASSWORDPOLICY_ADMIN_ALLOWUNBLOCK_LABEL' => 'Freigabe-Optionen',
-    'OXPS_PASSWORDPOLICY_ADMIN_ALLOWUNBLOCK' => 'Erlaube blockierten Benutzern ein selbständiges Aufheben der Blockierung durch Zurücksetzen des eigenen Passworts.',
     'OXPS_PASSWORDPOLICY_ADMIN_PASSWORDSTRENGTH' => 'Passwort Stärke Einstellungen',
     'OXPS_PASSWORDPOLICY_ADMIN_MINLENGTH' => 'Minimale Länge',
     'OXPS_PASSWORDPOLICY_ADMIN_MINLENGTH_HINT' => 'Min. erlaubte Passwortlänge (6 oder mehr).',
