+++ views/admin/en/passwordpolicy_lang.php
@@ -21,13 +21,6 @@ $aLang = array(
     'charset' => 'UTF-8',
     'oxpspasswordpolicy' => 'Password Policy',
     'OXPS_PASSWORDPOLICY_ADMIN_TITLE' => 'Password policy settings',
-    'OXPS_PASSWORDPOLICY_ADMIN_ACCOUNTBLOCK' => 'Automatic account blocking settings',
-    'OXPS_PASSWORDPOLICY_ADMIN_MAXATTEMPTS' => 'Maximum login attempts',
-    'OXPS_PASSWORDPOLICY_ADMIN_MAXATTEMPTS_HINT' => 'Maximum number of unsuccessful login attempts in a row before user account is blocked.',
-    'OXPS_PASSWORDPOLICY_ADMIN_TRACKINGPERIOD' => 'Tracking period',
-    'OXPS_PASSWORDPOLICY_ADMIN_TRACKINGPERIOD_HINT' => 'Time period in minutes to track unsuccessful login attempts sequence.',
-    'OXPS_PASSWORDPOLICY_ADMIN_ALLOWUNBLOCK_LABEL' => 'Unblock options',
-    'OXPS_PASSWORDPOLICY_ADMIN_ALLOWUNBLOCK' => 'Allow blocked users to unblock themselves on successful password reset.',
     'OXPS_PASSWORDPOLICY_ADMIN_PASSWORDSTRENGTH' => 'Password strength settings',
     'OXPS_PASSWORDPOLICY_ADMIN_MINLENGTH' => 'Minimal length',
     'OXPS_PASSWORDPOLICY_ADMIN_MINLENGTH_HINT' => 'Minimal allowed password length (6 or more).',
