+++ views/admin/it/passwordpolicy_lang.php
@@ -21,13 +21,6 @@ $aLang = array(
     'charset' => 'UTF-8',
     'oxpspasswordpolicy' => '[TR - Password Policy]',
     'OXPS_PASSWORDPOLICY_ADMIN_TITLE' => '[TR - Password policy settings]',
-    'OXPS_PASSWORDPOLICY_ADMIN_ACCOUNTBLOCK' => '[TR - Automatic account blocking settings]',
-    'OXPS_PASSWORDPOLICY_ADMIN_MAXATTEMPTS' => '[TR - Maximum login attempts]',
-    'OXPS_PASSWORDPOLICY_ADMIN_MAXATTEMPTS_HINT' => '[TR - Maximum number of unsuccessful login attempts in a row before user account is blocked.]',
-    'OXPS_PASSWORDPOLICY_ADMIN_TRACKINGPERIOD' => '[TR - Tracking period]',
-    'OXPS_PASSWORDPOLICY_ADMIN_TRACKINGPERIOD_HINT' => '[TR - Time period in minutes to track unsuccessful login attempts sequence.]',
-    'OXPS_PASSWORDPOLICY_ADMIN_ALLOWUNBLOCK_LABEL' => '[TR - Unblock options]',
-    'OXPS_PASSWORDPOLICY_ADMIN_ALLOWUNBLOCK' => '[TR - Allow blocked users to unblock themselves on successful password reset.]',
     'OXPS_PASSWORDPOLICY_ADMIN_PASSWORDSTRENGTH' => '[TR - Password strength settings]',
     'OXPS_PASSWORDPOLICY_ADMIN_MINLENGTH' => '[TR - Minimal length]',
     'OXPS_PASSWORDPOLICY_ADMIN_MINLENGTH_HINT' => '[TR - Minimal allowed password length (6 or more).]',
